refactor/move rules to separate file
